Add utility classes, add rounded navbar code
Add a custom Container to replace the standard NavigationView. It
contains a NaviagtionView but the lower corners are rounded, the
background color can be customized (which will extends itself to the top
of the screen, ignoring the safe area insets and has a large title
style that does not shrinks).

Add the `AppearanceManager` class to manage all the appearance-related
settings in the app. It has a hack to set the default navbar alpha
channel to 0, which is called in the AppDelegate.

Replace the default ContentView.

Add a `AlertPresenter` class that manages all the alerts that should be
shown by the application.