Add Counter detail view, add localizations, some fixes
Add counter detail view that currently only visualizes a message and can
distinguish between the case when it is opened to edit a counter or to
create one.

Add localizations for the `it` and `en` locales.

Fix the static rounded navigation bar in the main screen.

Changed the counters collection type in CountersManager from a
Set<CounterCore> to an Array<CounterCore>.

Add example array of counters in the CountersManager to be used in the
SwiftUI previews.