Add custom TextField, improve detail view, localizations and better color management logic
Add custom TextField to draw a toolbar on top of the keyboard for the
CounterDetail's TextFields to toggle the sign and close the keyboard.

Add the detail view form.

Add some localizations.

Changed the color management logic of the `Counter` objects, and in the
rest of the application. Changed some colors in the app.