Add Chekpoint details view, some related edits
Add a Checkpoint details view, add the list of checkpoints in the
CounterDetail view.

Change how the `isEditing` variable in the CounterDetail view is
computed to allow to be independent from the value of the provided
Counter instance, except in the init of the class.

Add a PlaySoundAction to be executed in a Checkpoint.

Add some localizations.