Complete Codable implementation, apply fix to circular references
Codable implementation is complete, need to check if the fix for
circular references is working as intended.

Add a parameter and an initializer for managing the case when a counter
is just a placeholder containg the id to allow other classes to identify
it and get the actual counter instance from the counters manager.